Changes to Dyninst to use the new thread information and events provided by ProcControlAPI
Modified PCThread and BPatch_thread so PCThread now encapsulates any
functionality required to extract thread information from a process instead of
BPatch_thread.

PCEventHandler now creates and destroys PCThreads using the thread events
provided by ProcControlAPI, agnostic of the events actually available on a
platform.

The RT library now uses thread ids instead of lwp ids when locating a tramp
guard for a thread.

Also, fixes a bug in the ProcControl testsuite component where file
descriptors weren't being closed -- this was causing tests to fail when
run in combination with Dyninst.

+++ dynutil/h/dyntypes.h
@@ -126,7 +126,7 @@ namespace Dyninst
    typedef int PID;
    typedef int PROC_HANDLE;
    typedef int LWP;
-   typedef int THR_ID;
+   typedef long THR_ID;
 
 #define NULL_PID     -1
 #define NULL_LWP     -1
